+++ src/main/java/dev/yatloaf/modkrowd/config/Predicate.java
@@ -1,38 +1,66 @@
 package dev.yatloaf.modkrowd.config;
 
+import dev.yatloaf.modkrowd.cubekrowd.common.CKColor;
 import dev.yatloaf.modkrowd.cubekrowd.subserver.CreativeSubserver;
 import dev.yatloaf.modkrowd.cubekrowd.subserver.CubeKrowdSubserver;
 import dev.yatloaf.modkrowd.cubekrowd.subserver.MissileWarsSubserver;
 import dev.yatloaf.modkrowd.cubekrowd.subserver.Subserver;
-import dev.yatloaf.modkrowd.util.Util;
 import net.minecraft.text.Style;
 import net.minecraft.text.Text;
-import net.minecraft.util.Formatting;
 
+import java.util.HashMap;
 import java.util.Map;
 
-public enum Predicate {
-    NEVER("never", Style.EMPTY.withFormatting(Formatting.RED), s -> false),
-    CUBEKROWD("cubekrowd", Style.EMPTY.withFormatting(Formatting.GOLD), s -> s instanceof CubeKrowdSubserver),
-    CREATIVE("creative", Style.EMPTY.withFormatting(Formatting.AQUA), s -> s instanceof CreativeSubserver),
-    MISSILEWARS("missilewars", Style.EMPTY.withFormatting(Formatting.WHITE), s -> s instanceof MissileWarsSubserver),
-    ALWAYS("always", Style.EMPTY.withFormatting(Formatting.GREEN), s -> true);
+public class Predicate {
+    // Maybe this class should be renamed
+
+    public static final Map<String, Predicate> FROM_ID = new HashMap<>();
+
+    public static final Predicate NEVER = new Predicate("never", CKColor.RED.style, Test::never);
+    public static final Predicate CUBEKROWD = new Predicate("cubekrowd", CKColor.GOLD.style, Test::cubeKrowd);
+    public static final Predicate CREATIVE = new Predicate("creative", CKColor.AQUA.style, Test::creative);
+    public static final Predicate MISSILEWARS = new Predicate("missilewars", CKColor.WHITE.style, Test::missileWars);
+    public static final Predicate ALWAYS = new Predicate("always", CKColor.GREEN.style, Test::always);
 
     public final String id;
     public final Text name;
     public final Text tooltip;
-    private final java.util.function.Predicate<Subserver> enabled;
+    public final Test test;
 
-    Predicate(String id, Style style, java.util.function.Predicate<Subserver> enabled) {
+    public Predicate(String id, Style style, Test test) {
         this.id = id;
         this.name = Text.translatable("modkrowd.config.predicate." + id).setStyle(style);
         this.tooltip = Text.translatable("modkrowd.config.predicate." + id + ".tooltip");
-        this.enabled = enabled;
+        this.test = test;
+        FROM_ID.put(id, this);
     }
 
-    public static final Map<String, Predicate> FROM_ID = Util.arrayToMap(values(), item -> item.id, item -> item);
+    public boolean enabled(Subserver subserver, int permissionLevel) {
+        return this.test.test(subserver, permissionLevel);
+    }
 
-    public boolean enabled(Subserver subserver) {
-        return this.enabled.test(subserver);
+    @FunctionalInterface
+    public interface Test {
+        boolean test(Subserver subserver, int permissionLevel);
+
+        static boolean never(Subserver subserver, int permissionLevel) {
+            return false;
+        }
+        static boolean cubeKrowd(Subserver subserver, int permissionLevel) {
+            return subserver instanceof CubeKrowdSubserver;
+        }
+        static boolean creative(Subserver subserver, int permissionLevel) {
+            return switch (subserver) {
+                case CreativeSubserver ignored -> true;
+                case CubeKrowdSubserver ignored -> false;
+                default -> permissionLevel >= 2;
+            };
+        }
+        static boolean missileWars(Subserver subserver, int permissionLevel) {
+            return subserver instanceof MissileWarsSubserver;
+        }
+        static boolean always(Subserver subserver, int permissionLevel) {
+            return true;
+        }
     }
 }