+++ src/main/java/dev/yatloaf/modkrowd/mixin/EntityAccessor.java
@@ -1,0 +1,12 @@
+package dev.yatloaf.modkrowd.mixin;
+
+import net.minecraft.entity.Entity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+@Mixin(Entity.class)
+public interface EntityAccessor {
+
+    @Invoker
+    int callGetPermissionLevel();
+}